(feat) re-organize to make processes cleaner
