+++ hummingbot/cli/hummingbot_application.py
@@ -47,14 +47,12 @@ from hummingbot.cli.errors import (
     ArgumentParserError
 )
 from hummingbot.cli.config.config_var import ConfigVar
-from hummingbot.cli.config.in_memory_config_map import (
-    in_memory_config_map,
+from hummingbot.cli.config.in_memory_config_map import in_memory_config_map
+from hummingbot.cli.config.global_config_map import global_config_map
+from hummingbot.cli.config.helpers import (
     get_strategy_config_map,
     write_config_to_yml,
     load_required_configs,
-)
-from hummingbot.cli.config.global_config_map import global_config_map
-from hummingbot.cli.config.helpers import (
     parse_cvar_value,
     copy_strategy_template,
     get_erc20_token_addresses,
